Web Campus Map Location Base System
This program was developed on aound June 2018. The system included three layers structure: client (HTML, JavaScript, CSS), server (Java, or Servlet) and database (PostGIS). Client end included login service which was connected to account database when checking the authorities. The map data was drew on 2016, when I was a sorphomore, by digitalizing remote sensing image of informatics campus of my university with ArcGIS, and the style of map was created by QGIS and posted on the server. In this program, all data were imported into database and called by the servlet server. Map system was based on OpenLayer, a map Javascript Library.